src/requirements.txt
certifi==2022.5.18.1
google-api-core==2.8.1
google-auth==2.6.6
google-cloud-bigquery==3.1.0
google-cloud-core==2.3.0
google-cloud-logging==3.1.1
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
python-dateutil==2.8.2
google-cloud-error-reporting==1.5.2